AnalyticaArena/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Database
motor==3.3.2
pymongo==4.6.1

# Data Processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
xlrd==2.0.1

# AI Integration
google-generativeai==0.3.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Report Generation
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
